Add the new settings of the screen.
Add the new settings of the screen, because of the height and the
width was to big for the laptops with the resolution 1366 x 768
pixels. The previous resolution caused to the cross in the right
upper corner of the game screen was beyond of the screen. This
maked impossible to quite from the game.